requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
numpy>=1.20.0
scipy>=1.6.0